Added new test cases. Modified scripts to work with the new tests.
The new tests are taken from Dmitry's "an-transformer", and they
are meant for benchmarking against other implementations of AN
encoding. However, the tests also work well for verification.

To make some of the tests work with the "AN encoder", parts of
the tests' source code had to be moved into external source files.
The motivation for this was to prevent the "AN encoder" from
being applied to said parts of the code. The easiest way to
achieve this was to replace those parts of the code with calls to
external functions. (We can think of these calls as library calls,
hence the folder name "mylibs".)

The various scripts had to be adapted to work with this new
structure of the test suite.

TODO: Design a test system for verification, i.e. checking that
AN-encoded code produces the same results as the original!

TODO: Design a system for automated benchmarking!