Added different implementations for sdiv/udiv and shr/ashr.
NOTE: sdiv and udiv have not been properly tested! Neither has it
been verified that sdiv/udiv are applied appropriately (depending
on the signed-ness of the arguments).

Whether we really need to distinguish between signed and unsigned
operations depends on whether we treat coded values as 'signed' or
'unsigned'. The currently favoured approach seems to be to think of
coded values as 'signed'. However, a clean resolution of this issue
is tightly coupled to the promotion of all integer values to 64bit
integers (which still needs to be implemented).