Added library functions for writing (intermediate) results from test cases to a (binary) file.
This new method of logging results is intended to supersede the
previous logging with hashes/checksums. When relying on hashes or
checksums to determine whether a program has executed correctly,
we might miss errors that do not affect the final value of the
checksum. (In practise, however, such errors are probably very,
very rare.)

CAUTION: It is still possible that (intermediate) results that are
written to a file are correct despite the fact that an error has
occurred. This is particularly likely if the granularity at which
intermediate results are written out is very coarse. (In the worst
case, e.g., only one final result is written out.)