Added a new patch for the BFI plugin. This patch serves to simplify the mask computation for injection of a 'TXT' fault.
More detailed description of the new patch:
Ultimately, for a 'TXT' fault, the mask is applied to a single
byte of the instruction opcode. Application of the mask is achieved
by an xor operation with the byte from the opcode. Hence the mask
should be truncated to a single byte.

The old patch (i.e. "fallthrough") was renamed.