Implemented proper parsing for files specifying encoding profiles.
The current implementation consists of a hand-written lexer and
recursive-descent parser. In the future, one may want to replace
either of these with an automatically generated lexer/parser.

The key point of the current implementation is that the class
'EncodingProfile' serves as the interface between the parser for
profile-files and the actual encoder.